logic extracted to api, listingCard made re-usable, all and potential matches now displaying correctly, refactoring
