+++ src/components/Listings.svelte
@@ -1,7 +1,7 @@
 <script>
   import { onSnapshot, collection } from "firebase/firestore";
   import { db, reseedListingsDatabase } from "../utils/api";
-import ListingCard from "./ListingCard.svelte";
+  import ListingCard from "./ListingCard.svelte";
 
   let listings = [];
 
@@ -11,7 +11,7 @@ import ListingCard from "./ListingCard.svelte";
       let listingArray = [];
       querySnapshot.forEach((listing) => {
         let listingData = { ...listing.data(), id: listing.id };
-        listingArray = [listingData, ...listingArray];
+        listingArray = [listingData, ...listingArray]; 
       });
       listings = listingArray;
     }
@@ -21,14 +21,14 @@ import ListingCard from "./ListingCard.svelte";
 <main>
   <h1>All listings ({listings.length})</h1>
   <p>Listings</p>
-  <ListingCard />
+  <ListingCard {listings} />
   <button
     on:click={(event) => {
       reseedListingsDatabase(event, listings);
     }}>Re-seed database</button
   >
 
-  <table style="width:100%">
+  <table>
     <tr>
       <th>id</th>
       <th>Title</th>
@@ -39,7 +39,7 @@ import ListingCard from "./ListingCard.svelte";
     </tr>
     {#each listings as listing}
       <tr
-        ><td>{listing.id.substring(0, 5)}...</td>
+        ><td style="width: 400px">{listing.id}...</td>
         <td>{listing.title}</td>
         <td>{listing.description}</td>
         <td>{listing.condition}</td>
