+++ src/components/PotentialMatches.svelte
@@ -12,6 +12,7 @@
   import { onMount } from "svelte";
   import App from "../../../firesvelt/src/App.svelte";
   import Listings from "./Listings.svelte";
+  import { queryPotentialUsers, queryPotentialMatchItems } from "../utils/api";
 
   let signedIn;
 
@@ -25,9 +26,7 @@
     return queryPotentialUsers(current_user).then((likingUsers) => {
       queryPotentialMatchItems(likingUsers)
         .then((items) => {
-          console.log(items, "ITEMS");
           potentialMatches = items;
-          return items;
         })
         .catch((error) => {
           const errorCode = error.code;
@@ -37,34 +36,6 @@
     });
   };
 
-  const queryPotentialUsers = async (current_user) => {
-    const query1 = query(
-      collection(db, "matches"),
-      where("item_owner_id", "==", current_user)
-    );
-    const querySnapshot = await getDocs(query1);
-    let usersThatLikedMyItem = [];
-    querySnapshot.forEach((doc) => {
-      let userData = { ...doc.data() };
-      usersThatLikedMyItem = [userData.liking_user_id, ...usersThatLikedMyItem];
-    });
-    return usersThatLikedMyItem;
-  };
-
-  const queryPotentialMatchItems = async (likingUsers) => {
-    const query2 = query(
-      collection(db, "listings"),
-      where("user_id", "in", likingUsers)
-    );
-    const querySnapshot = await getDocs(query2);
-    let potentialMatchItems = [];
-    querySnapshot.forEach((doc) => {
-      let itemData = { ...doc.data(), id: doc.id };
-      potentialMatchItems = [itemData, ...potentialMatchItems];
-    });
-    return potentialMatchItems;
-  };
-
   onMount(async () => {
     getPotMatches(`${signedIn.uid}`);
   });
@@ -72,13 +43,7 @@
 
 <main>
   <h2>Potential Matches</h2>
-
-  {#each potentialMatches as listings}
-    <p>{listings.title}</p>
-  {:else}
-    <p>Loading.App..</p>
-  {/each}
-  <!-- <ListingCard /> -->
+  <ListingCard listings={potentialMatches} />
 </main>
 
 <style>
